+++ tests/common/test_linear_operations_classes.cpp
@@ -100,7 +100,7 @@ void test_matrix3d_operations() {
     for (size_t i = 0; i < 2; ++i) {
         for (size_t j = 0; j < 3; ++j) {
             for (size_t k = 0; k < 4; ++k) {
-                mat3d(i, j, k) = i * 100 + j * 10 + k;
+                mat3d(i, j, k) = static_cast<double>(i * 100 + j * 10 + k);
             }
         }
     }
